Updated file names for components. Moved the console list used for the console button component into a seperate file.
