+++ client/src/components/GameResults/GameResults.js
@@ -1,7 +1,7 @@
 import { capitalizeWords, formatDate } from '../../utils/helpers';
 
 import React from 'react';
-import { ReviewStars } from '../ReviewStars';
+import { ReviewStars } from '../ReviewStars/ReviewStars';
 import { reviewTypes } from '../../utils/renderScore';
 import { useSearchContext } from '../../utils/context/SearchState';
