+++ client/src/constants/consoleList.js
@@ -1,0 +1,6 @@
+export const consoles = [
+	{ name: 'playstation', color: 'link' },
+	{ name: 'xbox', color: 'success' },
+	{ name: 'nintendo-switch', color: 'danger' },
+	{ name: 'steam', color: 'light' },
+];
